+++ lib/lightning/config/bootstrap.ex
@@ -662,6 +662,10 @@ defmodule Lightning.Config.Bootstrap do
     config :lightning, :ui_metrics_tracking,
       enabled: env!("UI_METRICS_ENABLED", &Utils.ensure_boolean/1, false)
 
+    config :lightning,
+           :broadcast_work_available,
+           env!("BROADCAST_WORK_AVAILABLE", &Utils.ensure_boolean/1, true)
+
     config :lightning, :book_demo_banner,
       enabled: false,
       calendly_url: nil,